#70 setup basic test and standards #63 remove unnecessary imports + fix namespaces + fixer + app -> text_extract_api
