+++ .gitignore
@@ -1,4 +1,4 @@
-app/__pycache__/*
+text_extract_api/__pycache__/*
 .env
 .env.local
 *.pyc
@@ -8,3 +8,5 @@ storage/*
 client_secret*.json    
 .env.localhost
 .idea
+# Python good practice ignore
+*.egg-info/
