--- app/Dockerfile
@@ -1,35 +1,0 @@
-# Use an official Python runtime as a parent image
-FROM python:3.10-slim
-
-
-# Set the working directory
-WORKDIR /app
-
-RUN echo 'Acquire::http::Pipeline-Depth 0;\nAcquire::http::No-Cache true;\nAcquire::BrokenProxy true;\n' > /etc/apt/apt.conf.d/99fixbadproxy
-
-# Clear the APT cache, update package lists, and install dependencies
-RUN apt-get clean && rm -rf /var/lib/apt/lists/* \
-    && apt-get update --fix-missing \
-    && apt-get install -y \
-        libgl1-mesa-glx \
-        tesseract-ocr \
-        libtesseract-dev \
-        poppler-utils \
-        libpoppler-cpp-dev \
-    && rm -rf /var/lib/apt/lists/*
-
-# Copy the requirements file into the container
-COPY requirements.txt .
-
-
-# Install any needed packages specified in requirements.txt
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Copy the rest of the application code
-COPY . .
-
-# Expose the port the FastAPI app runs on
-EXPOSE 8000
-
-# Define the command to run your application
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000"]
