--- app/__init__.py
@@ -1,4 +1,0 @@
-from . import extract
-from . import files
-
-__all__ = ["files", "extract"]
