--- app/extract/__init__.py
@@ -1,3 +1,0 @@
-from app import files
-
-__all__ = ["files"]
