--- app/files/__init__.py
@@ -1,5 +1,0 @@
-from .file_formats.file_format import FileFormat
-from .file_formats.image_file_format import ImageFileFormat
-from .file_formats.pdf_file_format import PdfFileFormat
-
-__all__ = ["FileFormat", "ImageFileFormat", "PdfFileFormat"]
