+++ pyproject.toml
@@ -1,0 +1,41 @@
+[build-system]
+requires = ["wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "text_extract_api"
+version = "0.1.0"
+description = "Images and documents (PDF, Word, PPTX ...) extraction and parse API using state of the art modern OCRs + Ollama supported models. Anonymize documents. Remove PII. Convert any document or picture to structured JSON or Markdown"
+authors = [
+]
+readme = "README.md"
+requires-python = ">=3.8"
+dependencies = [
+    "fastapi",
+    "uvicorn",
+    "celery",
+    "redis",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest",
+    "black",
+    "isort",
+    "flake8",
+]
+
+[tool.black]
+line-length = 88
+
+[tool.isort]
+profile = "black"
+multi_line_output = 3
+
+[tool.pytest]
+testpaths = ["tests"]
+python_files = ["test_*.py"]
+
+[tool.flake8]
+max-line-length = 88
+exclude = ["venv", "__pycache__", ".git", ".mypy_cache"]
